Changed so that citizenmodel is merged with citizen in one class. And I also deleted CitizenModel
